Adds starting code
