+++ package.json
@@ -32,6 +32,7 @@
     "react": "^16.6.0",
     "react-dom": "^16.6.0",
     "react-flippy": "^0.1.2",
+    "react-quill": "^1.3.2",
     "react-router-dom": "^4.3.1",
     "reactstrap": "^6.5.0",
     "sequelize": "^4.41.0",
